refactor: _predictive_utils and astrological_subject_factory
  _predictive_utils.py            -> predictive/utils.py
  astrological_subject_factory.py -> astrological_subject/factory.py

predictive/ drops the leading underscore: one underscored package among
forty clean ones is itself an inconsistency, and the module was never
private in practice -- it exports PTOLEMAIC_ASPECTS to the top-level
namespace and is imported by fifteen factories. What is internal is said
in the docstring and enforced by __all__, which exports only that one name.

astrological_subject/factory.py holds a single 3000-line class and is
moved as-is. Splitting a class means extracting methods or introducing
mixins -- a change to the code, not to where it lives -- and that does not
belong in a refactor whose contract is identical behaviour.
