+++ examples/modern_chart_john_lennon.py
@@ -1,4 +1,4 @@
-from kerykeion.astrological_subject_factory import AstrologicalSubjectFactory
+from kerykeion.astrological_subject.factory import AstrologicalSubjectFactory
 from kerykeion.chart_data.factory import ChartDataFactory
 from kerykeion import ChartDrawer
 from pathlib import Path
