+++ examples/synastry_data_example.py
@@ -1,4 +1,4 @@
-from kerykeion.astrological_subject_factory import AstrologicalSubjectFactory
+from kerykeion.astrological_subject.factory import AstrologicalSubjectFactory
 from kerykeion.chart_data.factory import ChartDataFactory
 
 # Create a natal chart data
