+++ kerykeion/composite_subject/factory.py
@@ -43,8 +43,8 @@ import logging
 from typing import Union
 
 # Fix the circular import by changing this import
-from kerykeion.astrological_subject_factory import AstrologicalSubjectFactory, _GEO_TOPO_PERSPECTIVES
-from kerykeion._predictive_utils import jd_to_ymd_hms
+from kerykeion.astrological_subject.factory import AstrologicalSubjectFactory, _GEO_TOPO_PERSPECTIVES
+from kerykeion.predictive.utils import jd_to_ymd_hms
 from kerykeion.schemas.exceptions import KerykeionException
 from kerykeion.schemas.models import CompositeSubjectModel, AstrologicalSubjectModel, PolarHouseFallbackModel
 from kerykeion.schemas.literals import (
