+++ kerykeion/planetary_phenomena/factory.py
@@ -18,7 +18,7 @@ import logging
 from typing import List, Optional
 
 from kerykeion.ephemeris_backend import ephe, ephemeris_session
-from kerykeion._predictive_utils import validate_julian_day
+from kerykeion.predictive.utils import validate_julian_day
 from kerykeion.schemas.exceptions import KerykeionException
 from kerykeion.settings.config_constants import POINT_NUMBER_MAP
