+++ kerykeion/report/generator.py
@@ -1541,7 +1541,7 @@ class ReportGenerator:
 
 
 if __name__ == "__main__":
-    from kerykeion.astrological_subject_factory import AstrologicalSubjectFactory
+    from kerykeion.astrological_subject.factory import AstrologicalSubjectFactory
     from kerykeion.chart_data.factory import ChartDataFactory
     from kerykeion.composite_subject.factory import CompositeSubjectFactory
     from kerykeion.planetary_returns.factory import PlanetaryReturnFactory
