+++ kerykeion/secondary_progressions/solar_arc.py
@@ -40,7 +40,7 @@ from kerykeion.aspects.orb_utils import (
 from kerykeion.schemas import KerykeionException
 from kerykeion.schemas.literals import SIGN_CODES
 from kerykeion.schemas.models import AstrologicalSubjectModel
-from kerykeion._predictive_utils import gather_active_points, build_aspect_settings, PTOLEMAIC_ASPECTS
+from kerykeion.predictive.utils import gather_active_points, build_aspect_settings, PTOLEMAIC_ASPECTS
 from kerykeion.utilities import _ZODIAC_SIGNS, get_planet_house, HOUSE_FIELD_NAMES
 
 from .factory import SecondaryProgressionFactory
