+++ kerykeion/sign_ingresses/factory.py
@@ -25,7 +25,7 @@ from typing import List, Literal, Optional
 
 from kerykeion.ephemeris_backend import ephe, ephemeris_session
 from kerykeion.settings.config_constants import POINT_NUMBER_MAP
-from kerykeion._predictive_utils import is_iso_date_only, jd_to_iso_utc as _jd_to_iso, validate_julian_bounds
+from kerykeion.predictive.utils import is_iso_date_only, jd_to_iso_utc as _jd_to_iso, validate_julian_bounds
 
 from kerykeion.schemas.exceptions import KerykeionException
 from kerykeion.schemas.literals import SiderealMode, ZodiacType
